+++ R/utils.R
@@ -5,3 +5,13 @@ make_call <- function(...) {
 vlapply <- function(X, FUN, ..., FUN.VALUE = logical(1)) {
   vapply(X, FUN, ..., FUN.VALUE = FUN.VALUE)
 }
+
+names2 <- function(x) {
+  nms <- names(x)
+  if (is.null(nms)) {
+    rep("", length(x))
+  } else {
+    nms[is.na(nms)] <- ""
+    nms
+  }
+}