+++ NEWS.md
@@ -1,5 +1,7 @@
 # 2.1.0.9000
 
+- `with_file()` function to automatically remove files.
+
 # 2.1.0
 
 - `with_connection()` function to automatically close R file connections.